Make gig creation a little more robust

+++ app/models/gig.rb
@@ -10,8 +10,7 @@ class Gig < ActiveRecord::Base
   belongs_to :act
   belongs_to :gig_entry
 
-  before_save :update_performances
-  after_save :update_gig_entry
+  before_save :prepare_for_save
   after_destroy :destroy_gig_entry
   
   default_scope { order(:first_performance) } 
@@ -22,20 +21,16 @@ class Gig < ActiveRecord::Base
   def self.find_by_entry(id)
     where(gig_entry_id: id).first || new 
   end
-  
-  def update_performances
-    performances.delete_if { |perf_date| self.dates.include? perf_date.to_s }
-                .each        &:mark_for_destruction
-    self.dates  .delete_if { |date| date_strings.include? date }
-                .each      { |date| performances.build date: date, gig_id: id }
-                
-    remaining = performances.delete_if(&:marked_for_destruction?).map(&:date)
-    self.first_performance = remaining.min
-    self.last_performance =  remaining.max
+
+  def prepare_for_save
+    perfs = updated_performances
+    self.first_performance = perfs.min
+    self.last_performance =  perfs.max
+    self.gig_entry = GigEntry.build! self
   end
   
   def date_range
-    min, max = performances.minimum(:date), performances.maximum(:date)
+    min, max = self.first_performance, self.last_performance
     return if min.blank? || max.blank?
     
     y, m, d =     min.year == max.year, min.month == max.month, min.day == max.day
@@ -55,17 +50,22 @@ class Gig < ActiveRecord::Base
   end
 
   private
+  
+  def updated_performances 
+    performances.delete_if { |perf_date| self.dates.include? perf_date.to_s }
+                .each        &:mark_for_destruction
+    dates       .delete_if { |date| date_strings.include? date }
+                .each      { |date| performances.build date: date, gig_id: id }
+                
+    performances.delete_if(&:marked_for_destruction?).map(&:date)
+  end
     
   def date_strings
     performances.map &:to_s
   end
-  
-  def update_gig_entry
-    self.gig_entry = GigEntry.build(self)
-  end
-  
+
   def destroy_gig_entry
-    self.gig_entry.destroy
+    self.gig_entry.destroy if self.gig_entry
   end
   
 end
